--- server/config/cloudinaryConfig.js
@@ -1,13 +1,0 @@
-import cloudinary from 'cloudinary';
-import dotenv from 'dotenv';
-
-dotenv.config();
-const cloudinaryConfig = (req, res, next) => {
-  cloudinary.config({
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY,
-    api_secret: process.env.CLOUDINARY_API_SECRET,
-  });
-  next();
-};
-export default cloudinaryConfig;
